Preference conversion (#28)
This PR updates the ConversionHelper and the locales.js files to the latest versions (no change in workflow) and it moves away from the legacy pref system.

It introduces the LegacyPrefs API, which allows to access the legacy pref system from the WX part of the addon. It is only used to migrate the old pref values into the WX storage system. This requires changes to manifest.json:
 * LegacyPrefs API is defined
 * preferences.js is loaded into background page, which allows us to set our default values during startup
 * storage permission is added

In background.js the following changes are required:
 * set default pref values
 * migrate pref values (can be removed at some later time)

The settings dialog (editemailsubjectSettings.xul) requires the following changes:
 * return to a preference attribute to identify DOM elements, that hold preferences
 * load the preference script, which includes some pref load and save methods
 * ConversionHelper does not need to be loaded anymore, the locales.js script and the preferences.js script load it on demand.
 * The script editemailsubjectSettings.js no longer accesses the legacy pref system, but registers the load() and save() methods of the preferences.js script

The edit dialog (editemailsubject.xul) requires the following changes:
 * ConversionHelper does not need to be loaded anymore, the locales.js script and the preferences.js script load it on demand.
 * The script editemailsubject.js no longer accesses the legacy pref system, but uses the getPrefValue() method of the preferences.js script